refactor(tier lists): the pumbility store joins ITierListRepository
A PUMBILITY tier list is a stored tier list — same entity, one extra dimension
(the cohort) and a count column — so its persistence belongs on the tier-list
family's repository, not on a port of its own. ITierListRepository gains the
three pumbility-list methods, EFTierListRepository absorbs the implementation,
and IPumbilityCensusRepository / EFPumbilityCensusRepository are deleted (port
and implementation together). The folder records move to Domain/Records so the
shared port can speak them.
