+++ ScoreTracker/ScoreTracker.ChartIntelligence/Infrastructure/EFTierListRepository.cs
@@ -142,5 +142,68 @@ namespace ScoreTracker.ChartIntelligence.Infrastructure
 
             foreach (var name in tierLists) _cache.Remove(TierListKey(mix, name));
         }
+
+        public async Task SavePumbilityTierLists(MixEnum mix, ChartType chartType, DifficultyLevel level,
+            IReadOnlyDictionary<string, PumbilityTierListFolder> byCohort,
+            CancellationToken cancellationToken)
+        {
+            await using var database = await _factory.CreateDbContextAsync(cancellationToken);
+            var mixId = MixIds.For(mix);
+            var typeName = chartType.ToString();
+            var levelInt = (int)level;
+            var existing = await database.Set<PumbilityCensusEntryEntity>()
+                .Where(e => e.MixId == mixId && e.ChartType == typeName && e.Level == levelInt)
+                .ToArrayAsync(cancellationToken);
+            database.Set<PumbilityCensusEntryEntity>().RemoveRange(existing);
+            foreach (var (cohortKey, folder) in byCohort)
+            foreach (var entry in folder.Entries)
+                await database.Set<PumbilityCensusEntryEntity>().AddAsync(new PumbilityCensusEntryEntity
+                {
+                    MixId = mixId,
+                    ChartType = typeName,
+                    Level = levelInt,
+                    CohortKey = cohortKey,
+                    ChartId = entry.ChartId,
+                    Appearances = entry.Appearances,
+                    CohortSize = folder.CohortSize,
+                    Category = entry.Category.ToString(),
+                    Order = entry.Order
+                }, cancellationToken);
+
+            await database.SaveChangesAsync(cancellationToken);
+        }
+
+        public async Task<PumbilityTierListFolder> GetPumbilityTierList(MixEnum mix, ChartType chartType,
+            DifficultyLevel level, string cohortKey, CancellationToken cancellationToken)
+        {
+            await using var database = await _factory.CreateDbContextAsync(cancellationToken);
+            var mixId = MixIds.For(mix);
+            var typeName = chartType.ToString();
+            var levelInt = (int)level;
+            var rows = await database.Set<PumbilityCensusEntryEntity>()
+                .Where(e => e.MixId == mixId && e.ChartType == typeName && e.Level == levelInt
+                            && e.CohortKey == cohortKey)
+                .ToArrayAsync(cancellationToken);
+            return new PumbilityTierListFolder(
+                rows.Select(e => new PumbilityTierListRecord(e.ChartId, e.Appearances,
+                    Enum.Parse<TierListCategory>(e.Category), e.Order)).ToArray(),
+                rows.Length == 0 ? 0 : rows[0].CohortSize);
+        }
+
+        public async Task<IEnumerable<(ChartType ChartType, int Level)>> GetPumbilityTierListFolders(
+            MixEnum mix, string cohortKey, CancellationToken cancellationToken)
+        {
+            await using var database = await _factory.CreateDbContextAsync(cancellationToken);
+            var mixId = MixIds.For(mix);
+            // A folder every one of whose rows reads zero is a folder this cohort cannot speak
+            // for — it is written, so the rebuild does not have to remember which folders it
+            // skipped, but it must not be offered.
+            return (await database.Set<PumbilityCensusEntryEntity>()
+                    .Where(e => e.MixId == mixId && e.CohortKey == cohortKey && e.Appearances > 0)
+                    .Select(e => new { e.ChartType, e.Level })
+                    .Distinct()
+                    .ToArrayAsync(cancellationToken))
+                .Select(e => (Enum.Parse<ChartType>(e.ChartType), e.Level));
+        }
     }
 }
