+++ ScoreTracker/ScoreTracker.Tests/ApplicationTests/BlendedTierListHandlerTests.cs
@@ -254,15 +254,15 @@ public sealed class BlendedTierListHandlerTests
             playerStats.Setup(p => p.GetStats(It.IsAny<MixEnum>(), It.IsAny<Guid>(), It.IsAny<CancellationToken>()))
                 .ReturnsAsync((MixEnum _, Guid id, CancellationToken _) => StatsFor(id));
         }
-        var census = new Mock<IPumbilityCensusRepository>();
-        census.Setup(c => c.GetFolder(It.IsAny<MixEnum>(), It.IsAny<ChartType>(), It.IsAny<DifficultyLevel>(),
-                It.IsAny<string>(), It.IsAny<CancellationToken>()))
-            .ReturnsAsync(new PumbilityCensusFolder(Array.Empty<PumbilityCensusRecord>(), 0));
+        var tierLists = new Mock<ITierListRepository>();
+        tierLists.Setup(c => c.GetPumbilityTierList(It.IsAny<MixEnum>(), It.IsAny<ChartType>(),
+                It.IsAny<DifficultyLevel>(), It.IsAny<string>(), It.IsAny<CancellationToken>()))
+            .ReturnsAsync(new PumbilityTierListFolder(Array.Empty<PumbilityTierListRecord>(), 0));
         // The real projector over the same stubbed ports: the Projection source is driven by
         // cohort membership and peer scores, which is what these fixtures already set up.
         return new BlendedTierListHandler(mediator.Object, charts.Object,
             new Mock<ICurrentUserAccessor>().Object, new MemoryCache(new MemoryCacheOptions()),
             new ScoreProjector(scores.Object, playerStats.Object, new Mock<IPlayerHistoryRepository>().Object),
-            census.Object, new Mock<ITitleRepository>().Object, playerStats.Object, scores.Object);
+            tierLists.Object, new Mock<ITitleRepository>().Object, playerStats.Object, scores.Object);
     }
 }
